+++ src/routes/index.ts
@@ -3,7 +3,9 @@ import { userController } from '../controllers/user.controller';
 
 export const userRouter = Router();
 
-userRouter.get('/', userController.getUsers.bind(userController));
-userRouter.post('/', json(), userController.createUser.bind(userController));
-userRouter.patch('/', json(), userController.updateUser.bind(userController));
-userRouter.delete('/', userController.deleteUser.bind(userController));
+userRouter
+  .get('/', userController.getUsers.bind(userController))
+  .get('/:id', userController.getUserById.bind(userController))
+  .post('/', json(), userController.createUser.bind(userController))
+  .patch('/', json(), userController.updateUser.bind(userController))
+  .delete('/', userController.deleteUser.bind(userController));
