+++ src/services/index.ts
@@ -1,2 +1,1 @@
 export * from './memory.user.service';
-export * from './user.service';
